requirements.txt
Pillow>=10.0.0
cryptography>=41.0.0
Flask>=3.0.0
gunicorn>=21.0.0
